day23: part2, still not working
Dijkstra is taking too long, and in some cases too much memory.

I found a couple of 'solutions' by hand, which were incorrect due to not respecting the rule about not moving in the corridor unless into home. I guess part 1 worked in spite of ignoring that rule :shrug:

On the other hand, the rule puts a constraint of possible states, which should significantly reduce the search space...